+++ bedrock/src/cli/bedrock-cli.js
@@ -16,15 +16,14 @@
     with Bedrock; if not, see <https://www.gnu.org/licenses>.
  */
 const program = require('commander');
-const { existsSync } = require('fs-extra');
 const { join, dirname } = require('path');
 const log = require('./log');
 const { bedrockEvents, EVENTS } = require('../server/events');
-const { Patterns } = require('../server/patterns');
 const { serve } = require('../server/server');
 const { version } = require('../../package.json');
 const { build, testPatternRenders } = require('./commands');
-const { processConfig, getMeta } = require('../lib/config');
+const { getMeta } = require('../lib/config');
+const { bootstrapFromConfigFile } = require('../lib/bootstrap');
 
 program
   .version(version)
@@ -48,52 +47,15 @@ const isDevMode = !dirname(__dirname)
 if (isDevMode) log.info('Bedrock Dev Mode on');
 
 const configPath = join(process.cwd(), 'bedrock.config.js');
-if (!existsSync(configPath)) {
-  log.error('Could not find bedrock.config.js file in CWD.');
-  process.exit(1);
-}
 
-const configPathBase = dirname(configPath);
-
-/** @type {BedrockConfig} */
-const config = processConfig(require(configPath), configPathBase);
-
-const patterns = new Patterns({
-  newPatternDir: config.newPatternDir,
-  patternPaths: config.patterns,
-  dataDir: config.data,
-  templateRenderers: config.templateRenderers,
-  assetSets: config.assetSets,
-  publicDir: config.public,
-  configPathBase,
-});
+const { patterns, config } = bootstrapFromConfigFile(configPath);
 
 const allTemplatePaths = patterns.getAllTemplatePaths();
-
-config.templateRenderers.forEach(templateRenderer => {
-  if (templateRenderer.init) {
-    templateRenderer.init({
-      config,
-      allPatterns: patterns.getPatterns(),
-      templatePaths: allTemplatePaths.filter(t => templateRenderer.test(t)),
-    });
-    log.info('Init done', null, `templateRenderer:${templateRenderer.id}`);
-  }
-});
-log.verbose('All templateRenderers init done');
-// const userPkgPath = join(process.cwd(), 'package.json');
-// let userPkg = {};
-// if (existsSync(userPkgPath)) userPkg = require(userPkgPath); // eslint-disable-line
-// const { scripts } = userPkg;
 
 program.command('serve').action(async () => {
   log.info('Serving...');
   const meta = await getMeta(config);
-  await serve({
-    config,
-    meta,
-    patterns,
-  });
+  await serve({ meta });
 });
 
 program.command('build').action(async () => {
@@ -116,11 +78,7 @@ program.command('start').action(async () => {
         ),
       }),
     ),
-    serve({
-      config,
-      meta,
-      patterns,
-    }),
+    serve({ meta }),
   ])
     .then(() => log.info('Started!', null, 'start'))
     .catch(err => {
@@ -135,6 +93,21 @@ program.command('test').action(async () => {
   const allPatterns = await patterns.getPatterns();
   await testPatternRenders(allPatterns, patterns);
   bedrockEvents.emit(EVENTS.SHUTDOWN);
+});
+
+program.command('demo-urls').action(() => {
+  const patternDemos = patterns.getPatternsDemoUrls();
+
+  patternDemos.forEach(patternDemo => {
+    patternDemo.templates.forEach(template => {
+      const urlList = template.demoUrls.join('\n');
+      console.log(`
+${patternDemo.id} - ${template.id}    
+${urlList}`);
+    });
+  });
+
+  process.exit(0);
 });
 
 program.parse(process.argv);