+++ examples/bootstrap/assets/design-tokens/animation.js
@@ -1,0 +1,21 @@
+module.exports = {
+  animate: {
+    duration: {
+      normal: {
+        value: '0.2s',
+      },
+      slow: {
+        value: '0.4s',
+      },
+    },
+    timing: {
+      linear: {
+        value: 'linear',
+        comment: 'This is the default browser setting',
+      },
+      ease: {
+        value: 'ease-in-out',
+      },
+    },
+  },
+};