+++ examples/bootstrap/assets/design-tokens/shadow.js
@@ -1,0 +1,18 @@
+module.exports = {
+  shadow: {
+    button: {
+      base: {
+        value: 'inset 0 1px 0 rgba( 255, 255, 255, .15), 0 1px 1px rgba( 0, 0, 0, .075)',
+      },
+      active: {
+        value: 'inset 0 3px 5px rgba( 0, 0, 0, .125)',
+      },
+    },
+    inner: {
+      value: 'inset 0 1px 2px rgba(34,34,34,25)',
+    },
+    dropdown: {
+      value: '0 .5rem 1rem rgba( 0, 0, 0, .175)'
+    }
+  },
+};